--- src/main/java/org/cloudfoundry/autosleep/util/Serializers.java
@@ -1,43 +1,0 @@
-package org.cloudfoundry.autosleep.util;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-
-import java.io.IOException;
-import java.time.Duration;
-import java.time.Instant;
-
-public class Serializers {
-
-    public static class IntervalSerializer extends JsonSerializer<Duration> {
-        @Override
-        public void serialize(Duration value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
-            gen.writeNumber(value.toMillis());
-        }
-    }
-
-    public static class IntervalDeserializer extends JsonDeserializer<Duration> {
-        @Override
-        public Duration deserialize(JsonParser parser, DeserializationContext ctx) throws IOException {
-            return Duration.ofMillis(parser.getLongValue());
-        }
-    }
-
-    public static class InstantSerializer extends JsonSerializer<Instant> {
-        @Override
-        public void serialize(Instant value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
-            gen.writeNumber(value.toEpochMilli());
-        }
-    }
-
-    public static class InstantDeserializer extends JsonDeserializer<Instant> {
-        @Override
-        public Instant deserialize(JsonParser parser, DeserializationContext ctx) throws IOException {
-            return Instant.ofEpochMilli(parser.getLongValue());
-        }
-    }
-}
