+++ src/main/java/org/cloudfoundry/autosleep/util/serializer/InstantDeserializer.java
@@ -1,0 +1,15 @@
+package org.cloudfoundry.autosleep.util.serializer;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+import java.io.IOException;
+import java.time.Instant;
+
+public class InstantDeserializer extends JsonDeserializer<Instant> {
+    @Override
+    public Instant deserialize(JsonParser parser, DeserializationContext ctx) throws IOException {
+        return Instant.ofEpochMilli(parser.getLongValue());
+    }
+}
