+++ src/main/java/org/cloudfoundry/autosleep/util/serializer/InstantSerializer.java
@@ -1,0 +1,16 @@
+package org.cloudfoundry.autosleep.util.serializer;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+import java.time.Instant;
+
+public class InstantSerializer extends JsonSerializer<Instant> {
+
+    @Override
+    public void serialize(Instant value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
+        gen.writeNumber(value.toEpochMilli());
+    }
+}
