+++ src/main/java/org/cloudfoundry/autosleep/util/serializer/IntervalDeserializer.java
@@ -1,0 +1,16 @@
+package org.cloudfoundry.autosleep.util.serializer;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+import java.io.IOException;
+import java.time.Duration;
+
+public class IntervalDeserializer extends JsonDeserializer<Duration> {
+
+    @Override
+    public Duration deserialize(JsonParser parser, DeserializationContext ctx) throws IOException {
+        return Duration.ofMillis(parser.getLongValue());
+    }
+}
