+++ src/main/java/org/cloudfoundry/autosleep/util/serializer/IntervalSerializer.java
@@ -1,0 +1,17 @@
+package org.cloudfoundry.autosleep.util.serializer;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+import java.time.Duration;
+
+public class IntervalSerializer extends JsonSerializer<Duration> {
+
+    @Override
+    public void serialize(Duration value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
+        gen.writeNumber(value.toMillis());
+    }
+
+}
